requirements.txt
blinker==1.9.0
click==8.3.0
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
python-dotenv==1.0.0
Werkzeug==3.1.3
